+++ src/cpx_io/cpx_system/cpx_ap/ap_module.py
@@ -361,10 +361,7 @@ class ApModule(CpxModule):
         # if available, read inputs
         values = []
         if self.input_channels:
-            # this needs to read all i/o register and decide later what to use
-            data = self.base.read_reg_data(
-                self.input_register, byte_input_size + byte_output_size
-            )
+            data = self.base.read_reg_data(self.input_register, byte_input_size)
 
             if self.apdd_information.product_category == ProductCategory.IO_LINK.value:
                 # IO-Link splits into byte_channel_size chunks. Assumes all channels are the same
@@ -385,6 +382,8 @@ class ApModule(CpxModule):
                 values.extend(bytes_to_boollist(data)[: len(self.input_channels)])
             elif all(c.data_type == "INT16" for c in self.input_channels):
                 values.extend(struct.unpack("<" + "h" * (len(data) // 2), data))
+            elif all(c.data_type == "UINT16" for c in self.input_channels):
+                values.extend(struct.unpack("<" + "H" * (len(data) // 2), data))
             else:
                 raise TypeError(
                     f"Input data type {self.input_channels[0].data_type} are not supported "
@@ -397,6 +396,10 @@ class ApModule(CpxModule):
 
             if all(c.data_type == "BOOL" for c in self.output_channels):
                 values.extend(bytes_to_boollist(data)[: len(self.output_channels)])
+            elif all(c.data_type == "INT16" for c in self.output_channels):
+                values.extend(struct.unpack("<" + "h" * (len(data) // 2), data))
+            elif all(c.data_type == "UINT16" for c in self.output_channels):
+                values.extend(struct.unpack("<" + "H" * (len(data) // 2), data))
             else:
                 raise TypeError(
                     f"Output data type {self.output_channels[0].data_type} are not supported "
@@ -468,12 +471,21 @@ class ApModule(CpxModule):
                 f"Data must be list of {len(self.output_channels)} elements"
             )
 
+        # Handle bool
         if all(c.data_type == "BOOL" for c in self.output_channels) and all(
             isinstance(d, bool) for d in data
         ):
             reg = boollist_to_bytes(data)
             self.base.write_reg_data(reg, self.output_register)
-            Logging.logger.info(f"{self.name}: Setting channels to {data}")
+            Logging.logger.info(f"{self.name}: Setting bool channels to {data}")
+            return
+
+        # Handle int
+        if all(
+            c.data_type in ["INT16", "UINT16"] for c in self.output_channels
+        ) and all(isinstance(d, int) for d in data):
+            for i, d in enumerate(data):
+                self.write_channel(i, d)
             return
 
         raise TypeError(
@@ -508,6 +520,7 @@ class ApModule(CpxModule):
             )
             return
 
+        # Handle bool
         if all(c.data_type == "BOOL" for c in self.output_channels) and isinstance(
             value, bool
         ):
@@ -515,7 +528,27 @@ class ApModule(CpxModule):
             data[channel] = value
             reg_content = boollist_to_bytes(data)
             self.base.write_reg_data(reg_content, self.output_register)
-            Logging.logger.info(f"{self.name}: Setting channel {channel} to {value}")
+            Logging.logger.info(
+                f"{self.name}: Setting bool channel {channel} to {value}"
+            )
+            return
+
+        # Handle int16
+        if self.output_channels[channel].data_type == "INT16" and isinstance(
+            value, int
+        ):
+            reg = struct.pack("<h", value)
+            self.base.write_reg_data(reg, self.output_register)
+            Logging.logger.info(f"{self.name}: Setting int channel to {value}")
+            return
+
+        # Handle uint16
+        if self.output_channels[channel].data_type == "UINT16" and isinstance(
+            value, int
+        ):
+            reg = struct.pack("<H", value)
+            self.base.write_reg_data(reg, self.output_register)
+            Logging.logger.info(f"{self.name}: Setting uint channel to {value}")
             return
 
         raise TypeError(